para rodar

+++ 007-gerenciamento-de-memoria/ex01/programa.cpp
@@ -10,7 +10,7 @@ class vector{
         {
             this->capacity_ += this->increase_size_;
             int *new_data = new int[this->capacity_];
-            for (int i = 0; i < this->size_; ++i)
+            for (unsigned int i = 0; i < this->size_ ; ++i)
                 new_data[i] = data[i];
             delete[] this->data;
             this->data = new_data;
@@ -32,7 +32,7 @@ class vector{
         }
         int get_at(int index)
         {
-            if (index < 0 or index >= size())
+            if (index < 0 or index >= (long long) size())
                 throw std::out_of_range("Index out of range");
             return this->data[index];
         }
